Rework vscode-specific command client code
