+++ apps/vscode/command_client/vscode.py
@@ -1,17 +1,23 @@
-from pickle import FALSE
 from typing import Any
 
 from talon import Context, Module, actions
 
 from .command_client import NoFileServerException, NotSet, run_command
 
+mod = Module()
+
 ctx = Context()
+linux_ctx = Context()
 
 ctx.matches = r"""
 app: vscode
 """
+linux_ctx.matches = r"""
+os: linux
+app: vscode
+"""
+
 ctx.tags = ["user.command_client"]
-mod = Module()
 
 
 def command_server_or_client_fallback(command_id: str, wait: bool):
@@ -90,3 +96,11 @@ class Actions:
         return actions.user.run_rpc_command_get(
             command_id, arg1, arg2, arg3, arg4, arg5
         )
+
+
+@linux_ctx.action_class("user")
+class LinuxUserActions:
+    def trigger_command_server_command_execution():
+        # Work around bug with upper f-keys in VSCode on Linux. See
+        # https://github.com/pokey/command-server/issues/9#issuecomment-963733930
+        actions.key("ctrl-shift-alt-p")
